% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_ExUtil_portfolio.R
\name{Class_MeanVar_portfolio}
\alias{Class_MeanVar_portfolio}
\title{S3 class MeanVar_portfolio}
\description{
Class MeanVar_portfolio is designed to construct mean-variance portfolios
with provided estimators of the mean vector, covariance matrix, and
inverse covariance matrix. It includes the following elements:
}
\section{Slots}{
\tabular{ll}{
   Element \tab Description \cr
   call \tab the function call with which it was created \cr
   cov_mtrx \tab the sample covariance matrix of the asset returns \cr
   inv_cov_mtrx \tab the inverse of the sample covariance matrix \cr
   means \tab sample mean vector of the asset returns \cr
   weights \tab portfolio weights \cr
   Port_Var \tab portfolio variance \cr
   Port_mean_return \tab expected portfolio return \cr
   Sharpe \tab portfolio Sharpe ratio \cr
}
}

\seealso{
summary.MeanVar_portfolio summary method for the class,
\link{new_MeanVar_portfolio} class constructor,
\link{validate_MeanVar_portfolio} class validator,
\link{MeanVar_portfolio} class helper.
}
