% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SP_daily_asset_returns}
\alias{SP_daily_asset_returns}
\title{Daily  log-returns  of selected constituents S&P500.}
\format{
a matrix with the first column containing the data and company names as column labels.
}
\source{
Yahoo finance
}
\usage{
SP_daily_asset_returns
}
\description{
Daily log-returns of selected constituents of S&P500 in percents. The data are sampled
in business time, i.e., weekends and holidays are omitted.
}
\keyword{datasets}
